Issue #4 Support pip distribution
- Restructure file locations to support pip distribution
- Update files to support pip packaging
