+++ src/regfmt/Constants.py
