+++ src/regfmt/Endian.py
