+++ src/regfmt/stairleftlayout.py
@@ -1,16 +1,15 @@
-from regfmtlib.svggeometry import *
-from regfmtlib import Register
+from regfmt.svggeometry import *
+from regfmt import Register
 
-
-def writeFieldNameCenterSVG(registerDB, topGroup, styleSheet, font, registerSpacing=5):
+def writeStairLeftSVG(registerDB, topGroup, styleSheet, font, registerSpacing=5):
     bitFieldSize = getBitFieldSize(font=font)
 
     registers: [Register] = registerDB.registers
     parentOrigin = Point()
     parentSize = Size(height=(bitFieldSize.height + registerSpacing) * len(registers))
 
+    maxFieldDisplacementYs = []
     displacementY = 0
-
     for register in registers:
         registerGroup = Group()
         topGroup.append(registerGroup)
@@ -35,6 +34,17 @@ def writeFieldNameCenterSVG(registerDB, topGroup, styleSheet, font, registerSpac
 
         displacementX = 0
 
+        fieldNames = list(filter(lambda x: x is not None, [field.name for field in register.fields]))
+        fieldCount = len(fieldNames)
+
+        fieldFrames = tuple(map(lambda x: getTextFrame(text=x, font=styleSheet.fieldName.truetype()), fieldNames))
+        fieldHeights = tuple(map(lambda x: x.size.height, fieldFrames))
+        maxFieldHeight = max(fieldHeights)
+
+        fieldDisplacementY = (maxFieldHeight + 10.0) * fieldCount
+        maxFieldDisplacementYs.append(fieldDisplacementY)
+
+        counter = 0
         for field in register.fields:
             fieldWidth = field.width * bitFieldSize.width
             fieldHeight = registerHeight
@@ -49,18 +59,6 @@ def writeFieldNameCenterSVG(registerDB, topGroup, styleSheet, font, registerSpac
                                       width=fieldWidth,
                                       height=fieldHeight,
                                       style=styleSheet.field))
-
-            if field.name:
-                fieldNameFrame = getTextFrame(text=field.name, font=styleSheet.fieldName.truetype())
-
-                fieldNameX = (fieldWidth / 2.0) + displacementX
-                fieldNameY = fieldNameFrame.origin.y + (fieldHeight / 2.0) - (fieldNameFrame.size.height / 2.0)
-
-                registerGroup.append(Text(text=field.name,
-                                          x=fieldNameX,
-                                          y=fieldNameY,
-                                          textAnchor='middle',
-                                          style=styleSheet.fieldName))
 
             if field.width > 1:
                 registerGroup.append(Text(str(field.leftIndex),
@@ -84,11 +82,40 @@ def writeFieldNameCenterSVG(registerDB, topGroup, styleSheet, font, registerSpac
                                           style=styleSheet.fieldIndex
                                           ))
 
+            if field.name:
+                fieldNameFrame = fieldFrames[counter]
+
+                fieldNameX = (fieldWidth / 2.0) + displacementX + 3.0
+                fieldNameY = fieldHeight + fieldNameFrame.origin.y + fieldDisplacementY + 10.0
+
+                registerGroup.append(Text(text=field.name,
+                                          x=fieldNameX,
+                                          y=fieldNameY,
+                                          textAnchor='left',
+                                          style=styleSheet.fieldName))
+
+                registerGroup.append(Line(x1=displacementX + (fieldWidth / 2.0),
+                                          y1=registerHeight,
+                                          x2=displacementX + (fieldWidth / 2.0 ),
+                                          y2=fieldNameY,
+                                          style=styleSheet.fieldNameLine))
+
+                fieldDisplacementY -= (maxFieldHeight + 10)
+
+                if parentSize.height < fieldNameY:
+                    parentSize.height = fieldNameY
+
+                counter += 1
+
             displacementX += fieldWidth
 
+    counter2 = 0
     for group in topGroup:
         group.translate(0, displacementY)
-        displacementY += (bitFieldSize.height + registerSpacing)
+        displacementY += (bitFieldSize.height * 2.0 + registerSpacing + maxFieldDisplacementYs[counter2])
+        if parentSize.height < displacementY:
+            parentSize.height = displacementY
+        counter2 += 1
 
     return (parentOrigin, parentSize)
