+++ src/regfmt/tests/test_endian.py
